VM execution tracing, debugging improvements
Misc. debugging improvements and fixes
======================================

- Set compile time options to use the default values only if they are not already defined, so that options can be defined externally, in a Makefile, for example. Previously, most compile-time options were hard-coded and had to be changed in the source code.
- Add some missing checks of return values from esc_string() and Parser_def_name(); output something appropriate to situation if NULL/0, like "<NULL>" or "<UNDEFINED>"
- Add ByteCode_def_name() to mirror Parser_def_name()
- Improve void ByteCode_print_instructions(ByteCode *self) to print symbolic values and strings referenced

Note: op_name() should be used to get an opcode's symbolic name, and the functions ByteCode_def_name and Parser_def_name() should be used to get a definition/identifier name (remember to check return value), rather than referencing internal structures directly; this prevents accessing out of range indexes.

Invalid Identifier fix / work-around
====================================

The parser VM will now return INVALID_IDENTIFIER (-11) if an invalid identifier (-1) is detected in an IDENT instruction. This is a run-time check for a situation that should not occur in the first place; byte-code should not include invalid references. The compiler needs a fix to return an error instead of silently generating byte code that includes such references. In this case, the invalid definition id of -1 is placed into an IDENT instruction argument when a definition cannot be found at compile time.

Tracing
=======

This commit adds some tracing options to the parser VM to help debug issues, or for use as a development aid. Tracing support is not compiled in by default; it has to be enabled by defining compile-time options. All tracing is off by default at run-time, and must be enabled in the parser instance before running Parser_parse().

VM_TRACE
--------

When VM_TRACE is defined as non-zero at compile time, and parser->vm_trace is set to non-zero, tracing information will be output to stderr *before* each instruction is executed. Info printed is as follows:

=     CNT   OFFSET       PC           OP   ARG

CNT: a running count of instructions executed (`cnt`)
OFFSET: current value of `offset` (read position within input)
PC: the value of `pc`, the VM program counter (current instruction in instruction table)
OP: current OP code in symbolic form (symbolic value of `op`)
ARG: argument of current instruction (an int value, `arg`)

The final column displays what `arg` refers to, which is instruction specific. If `arg` refers to an identifier, it will print the identifier name. If `arg` refers to a literal, or a char class being matching, it displays the string value. If `arg` is unused or numeric, a '-' is displayed in this column (see ARG column for the numeric value). To give the user a better idea where the VM is in the parsing process, this column is indented by tree_top, equivalent to the current tree stack depth.

Each line of the trace output reflects the state of the VM *before* the current instruction is executed.

Note: In the current implementation, instruction addresses in arguments of instructions, such as GOTO, are pre-subtracted by one, because pc is incremented unconditionally at the end of the VM loop. This can be a bit confusing when manually following traces and referencing the instruction table - you have to mentally add one to instruction addresses when referencing the instruction table.

To enable parser VM execution tracing:
1) Define VM_TRACE to a non-zero value to compile in the tracing code.
2) Before running Parser_parse(parser, ...), set parser->vm_trace to a non-zero value to enable tracing, or set to zero to disable tracing.

VM_PRINT_TREE
-------------

This compile-time option adds support to print the parse tree as it is built in Parser_parse().

To enable:
1) Define VM_PRINT_TREE to a non-zero value to compile in support
2) Before running Parser_parse(parse, ...), set parser->vm_print_tree to a non-zero value to enable printing, or set to zero to disable.

The parse tree is printed *after* an instruction is executed. To reduce redundant output, it is printed only if the tree was changed by the instruction just executed.

Implemention notes:
- The parse tree is built with child nodes being pushed in reverse order, and then fixed later during the 'unwrap' process after parsing has finished. The reason for this: optimization. The Node structure is optimized to require minimal pointer setting to build a tree. There are only two Node pointers in a Node: `head`, which points at the first child; and `next`, which points to the next sibling node. Children can be easily be pushed at the beginning of the list; to push them at the end would require iterating all the way to the end since there is no 'last child' pointer. Much of the tree building may end up being thown away due to backtracking, so it makes sense to minimize cycles used to build tree nodes, and then clean up the final tree when a parse is complete.
- When altering any code in the VM loop, you need to set tree_changed=1 whenever you change the parse tree. Wrap this in a check for VM_PRINT_TREE:
```
    #if VM_PRINT_TREE
        tree_changed = 1
    #endif
```
- The tree is printed using existing Parse_print_tree(), which prints to stdout. This really should go to stderr since VM_TRACE output goes there. TODO