Better implementation of TRIM ('<' ... '>')
TRIM acts much like a capture, but is better thought of as a token
boundary operator. The implementation simply 'trims' off the left or
right hand side of a match by adjusting the Node's offset and length.

Syntactically, '<'...'>' is like '('...')' grouping. Although many may
occur in a rule, only one TRIM will take effect. If nested, the
outermost level will be used. If multiple '<','>' pairs occur in
Sequence, the last encountered pair will be used. If '<','>' pairs occur
in parallel in a Choice, the first matched one will be used.