Change offset/length to always contain token
New behavior for ChpegNode offset/length fields:

        size_t offset; // full match offset (without TRIM) / token offset (with TRIM)
        size_t length; // full match length (without TRIM) / token length (with TRIM)

    #if CHPEG_EXTENSION_TRIM
        size_t match_offset; // full match offset
        size_t match_length; // full match length
    #endif

At one point things were done this way (but only match_length, for
packrat), then it was changed so there was a separate token
offset/length for TRIM. This simplifies code some, but forces the user
to get the token offset/length they're interested in from different
places depending on TRIM extension being compiled in or not.

The downside of this change: things that require the full match, like
packrat and s2 simplification, have to work out where to get the full
match offset/length from.