Fix packrat with -s2 simplification
The new (-s2) simplification algorithm requires non-zero-length ignored
nodes to be present in the parse tree (these are removed at a later
stage). This fix corrects the case where such nodes, obtained from
packrat cache, were not being added to the parse tree.

Prior to this fix, ignored nodes retrieved from the packrat cache were
not added to the parse tree, which caused the s2 algorithm to not
simplify these cases.