Stack-based TRIM implementation
With this change, TRIM does not require a 'current node' for storage.
This should fix issues, such as the wrong TRIM taking effect when
multiple levels of TRIM are used, while tree building is inhibited (it
will be inhibited by default in current implementation, due to the TRIM
auto-`{L}` feature).